requirements.txt
python-docx
PyPDF2
extract-msg
requests
faiss-cpu
sentence-transformers
python-dotenv